+++ staticx/api.py
@@ -156,13 +156,22 @@ class StaticxGenerator:
             raise LibExistsError(libname)
 
         # Copy the library to the temp dir before stripping
+        # TODO: Do this lazily if stripping, or if patchelf-ing
         tmplib = os.path.join(self.tmpdir, basename(libpath))
         logging.info("Copying {} to {}".format(libpath, tmplib))
         shutil.copy(libpath, tmplib)
 
+        # Audit library to check for problems
+        try:
+            self.check_library_rpath(libpath)
+        except (UnsupportedRpathError, UnsupportedRunpathError):
+            # Fix it by removing
+            logging.info("Removing RPATH/RUNPATH from library {}".format(tmplib))
+            remove_rpath(tmplib)
+
         if self.strip:
             # Strip the library
-            logging.info("Stripping binary {}".format(tmplib))
+            logging.info("Stripping library {}".format(tmplib))
             strip_elf(tmplib)
 
             libpath = tmplib
@@ -193,6 +202,38 @@ class StaticxGenerator:
                     " library {} present".format(libname, arcname))
         self._added_libs[arcname] = linklib
 
+
+    def check_library_rpath(self, path, dangerous_only=False):
+        """Inspect a library to see if it uses problematic RPATH/RUNPATH
+
+        See https://github.com/JonathonReinhart/staticx/issues/172
+        """
+        def is_dangerous(rpath):
+            # rpath can be:
+            # * Absolutute                  dangerous
+            # * Relative (to working dir)   dangerous (and stupid)
+            # * Relative (to $ORIGIN)       safe (as long as no ..)
+            if not rpath.startswith('$ORIGIN'):
+                return True
+
+            # There might be some odd corner cases here, but this
+            # conservative approach should be good enough.
+            if '..' in rpath:
+                return True
+
+            return False
+
+        with open_elf(path) as elf:
+            # Check for RPATH
+            rp = elf.get_rpath()
+            if rp and not dangerous_only and is_dangerous(rp.rpath):
+                raise UnsupportedRpathError(path, rp.rpath)
+
+            # Check for RUNPATH
+            rp = elf.get_runpath()
+            if rp:
+                # RUNPATH is always dangerous because it kills RPATH
+                raise UnsupportedRunpathError(path, rp.runpath)