+++ test/pyinstall-lib-runpath/shlib.c
@@ -1,0 +1,6 @@
+#include "shlib.h"
+
+int foo(void)
+{
+    return 42;
+}